students/K33392/laboratory_works/stukalov_artem/laboratory_work_3/pyproject.toml
[tool.poetry]
name = "laboratory-work-3"
version = "0.1.0"
description = ""
authors = ["TUTOR03 <stukalov.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
